undefined check

+++ routes/api/post/post.controller.js
@@ -10,62 +10,66 @@ exports.makePost = (req, res) => {
 	const d = new Date();
 	d.setUTCHours(d.getUTCHours() + 9);
 	const { title, content, base64, tags } = req.body;
-	const picKey = d.getFullYear()+'_'
-									+d.getMonth()+'_'
-									+d.getDate()+'_'
-									+d.getTime()+'_'
-									+d.getSeconds()+'_'
-									+req.decoded._id+'.jpg';
-	const picUrl = 'https://s3.amazonaws.com/fashionpobucket/' + picKey;
+	if (title === undefined || content === undefined || base64 === undefined || tags === undefined) {
+		return res.status(406).json({ message: 'parameter wrong' });
+	} else {
+		const picKey = d.getFullYear() + '_'
+			+ d.getMonth() + '_'
+			+ d.getDate() + '_'
+			+ d.getTime() + '_'
+			+ d.getSeconds() + '_'
+			+ req.decoded._id + '.jpg';
+		const picUrl = 'https://s3.amazonaws.com/fashionpobucket/' + picKey;
 
-	User.findOne({ _id : req.decoded._id }, function(err, user) {
-		if (err) return res.status(500).json({ error: err });
-		if (!user) return res.status(404).json({ message:'no such user' });
-		// console.log(base64);
-		let post = new Post({
-			title,
-			content,
-			tags,
-			picUrl,
-			writtenBy: req.decoded._id,
-			writtenAt: d
-		});
-		post.save(function(err, post) {
-			if (err) return res.status(500).json({ error:err });
-			let buf = new Buffer(base64.replace(/^data:image\/\w+;base64,/, ''), 'base64');
-			user.myPost.push(post._id);
-			user.save((err) => {
+		User.findOne({ _id: req.decoded._id }, function (err, user) {
+			if (err) return res.status(500).json({ error: err });
+			if (!user) return res.status(404).json({ message: 'no such user' });
+			// console.log(base64);
+			let post = new Post({
+				title,
+				content,
+				tags,
+				picUrl,
+				writtenBy: req.decoded._id,
+				writtenAt: d
+			});
+			post.save( function(err, post) {
 				if (err) return res.status(500).json({ error: err });
-				s3.putObject({ 
-					Bucket: 'fashionpobucket', 
-					Key: picKey,
-					Body: buf, 
-					ACL: 'public-read' 
-				}, function(err) {
-					if (err) {
-						return res.send({ message: err });
-					} else {
-						client
-							.labelDetection(picUrl)
-							.then(results => {
-								const labels = results[0].labelAnnotations;
-								let labelArray = [];
-								labels.forEach(label => labelArray.push(label.description));
-								return labelArray;
-							})
-							.then((array) => {
-								post.imageTags = array;
-								return post.save();
-							})
-							.catch(err => {
-								console.error('ERROR:', err);
-							});
-						return res.send({ message: 'upload success' });
-					}
+				let buf = new Buffer(base64.replace(/^data:image\/\w+;base64,/, ''), 'base64');
+				user.myPost.push(post._id);
+				user.save((err) => {
+					if (err) return res.status(500).json({ error: err });
+					s3.putObject({
+						Bucket: 'fashionpobucket',
+						Key: picKey,
+						Body: buf,
+						ACL: 'public-read'
+					}, function(err) {
+						if (err) {
+							return res.send({ message: err });
+						} else {
+							client
+								.labelDetection(picUrl)
+								.then(results => {
+									const labels = results[0].labelAnnotations;
+									let labelArray = [];
+									labels.forEach(label => labelArray.push(label.description));
+									return labelArray;
+								})
+								.then((array) => {
+									post.imageTags = array;
+									return post.save();
+								})
+								.catch(err => {
+									console.error('ERROR:', err);
+								});
+							return res.send({ message: 'upload success' });
+						}
+					});
 				});
 			});
 		});
-	});
+	}
 };
 
 exports.getAllPosts = (req, res) => {
@@ -116,31 +120,39 @@ exports.getMyPost = (req, res) => {
 };
 exports.deletePost = (req, res) => {
 	const { post_id } = req.params;
-	Post.deleteOne({ _id : post_id }, (err) => {
-		if (err) return res.status(500).json({ error: err });
-		return res.status(200).json({ 
-			message: 'post deleted successfully' 
+	if (post_id === undefined) {
+		return res.status(406).json({ message: 'parameter wrong' });
+	} else {
+		Post.deleteOne({ _id: post_id }, (err) => {
+			if (err) return res.status(500).json({ error: err });
+			return res.status(200).json({
+				message: 'post deleted successfully'
+			});
 		});
-	});
+	}
 };
 
 exports.updatePost = (req, res) => {
 	const { post_id } = req.params;
 	const { title, content } = req.body;
-	Post.findOne({ _id : post_id }, (err, post) => {
-		if (err) return res.status(500).json({ error: err });
-		if (!post) return res.status(404).json({ message:'no such post' });
-
-		post.title = title;
-		post.content = content;
-
-		post.save((err) => {
+	if (title === undefined || content === undefined) {
+		return res.status(406).json({ message: 'parameter wrong' });
+	} else {
+		Post.findOne({ _id: post_id }, (err, post) => {
 			if (err) return res.status(500).json({ error: err });
-			return res.status(200).json({ 
-				message: 'post updated successfully' 
+			if (!post) return res.status(404).json({ message: 'no such post' });
+
+			post.title = title;
+			post.content = content;
+
+			post.save((err) => {
+				if (err) return res.status(500).json({ error: err });
+				return res.status(200).json({
+					message: 'post updated successfully'
+				});
 			});
 		});
-	});
+	}
 };
 exports.getMyLikePost = (req, res) => {
 	User.findOne({ _id: req.decoded._id }, (err, user) => {
@@ -158,127 +170,150 @@ exports.getMyLikePost = (req, res) => {
 
 exports.likePost = (req, res) => {
 	const { post_id } = req.params;
-	Post.findOne({ _id: post_id })
-		.then((post) => {
-			post.likeCnt++;
-			post.save( (err, newPost) => {
-				User.findOne({ _id: req.decoded._id })
-					.then((user) => {
-						user.likePost.push(newPost);
-						user.save((err) => {
-							if (err) return res.status(500).json({ error: err });
-							return res.status(200).json({
-								message: 'post liked successfully'
+	if (post_id === undefined) {
+		return res.status(406).json({ message: 'parameter wrong' });
+	} else {
+		Post.findOne({ _id: post_id })
+			.then((post) => {
+				post.likeCnt++;
+				post.save( (err, newPost) => {
+					User.findOne({ _id: req.decoded._id })
+						.then((user) => {
+							user.likePost.push(newPost);
+							user.save((err) => {
+								if (err) return res.status(500).json({ error: err });
+								return res.status(200).json({
+									message: 'post liked successfully'
+								});
 							});
 						});
-					});
+				});
+			})
+			.catch((err) => {
+				return res.status(500).json({ error: err });
 			});
-		})
-		.catch((err) => {
-			return res.status(500).json({ error: err });
-		});
+	}
 };
 
 exports.deleteLike = (req, res) => {
 	const { post_id } = req.params;
-	Post.findOne({ _id: post_id }, (err, post) => {
-		if (err) return res.status(500).json({ error: err });
-		if (!post) return res.status(404).json({ message: 'no such post' });
-		post.likeCnt--;
-		post.save((err) => {
+	if (post_id === undefined) {
+		return res.status(406).json({ message: 'parameter wrong' });
+	} else {
+		Post.findOne({ _id: post_id }, (err, post) => {
 			if (err) return res.status(500).json({ error: err });
-			User.findOne({ _id: req.decoded._id }, (err, user) => {
+			if (!post) return res.status(404).json({ message: 'no such post' });
+			post.likeCnt--;
+			post.save((err) => {
 				if (err) return res.status(500).json({ error: err });
-				if (!user) return res.status(404).json({ message: 'no such user' });
-				let index = user.likePost.indexOf(post_id);
-				user.likePost.splice(index, 1);
-				user.save((err) => {
+				User.findOne({ _id: req.decoded._id }, (err, user) => {
 					if (err) return res.status(500).json({ error: err });
-					return res.status(200).json({ message: 
-						'post liked deleted successfully' 
+					if (!user) return res.status(404).json({ message: 'no such user' });
+					let index = user.likePost.indexOf(post_id);
+					user.likePost.splice(index, 1);
+					user.save((err) => {
+						if (err) return res.status(500).json({ error: err });
+						return res.status(200).json({ message: 
+							'post liked deleted successfully' 
+						});
 					});
 				});
 			});
 		});
-	});
+	}
 };
 
 exports.getPostById = (req, res) => {
 	const { post_id } = req.params;
-	Post.findOne({ _id: post_id })
-		.then((post) => {
-			return res.status(200).json({ post: post });
-		})
-		.catch((err) => {
-			return res.status(500).json({ error: err });
-		});
+	if (post_id === undefined) {
+		return res.status(406).json({ message: 'parameter wrong' });
+	} else {
+		Post.findOne({ _id: post_id })
+			.then((post) => {
+				return res.status(200).json({ post: post });
+			})
+			.catch((err) => {
+				return res.status(500).json({ error: err });
+			});
+	}
 };
 
 exports.viewPost = (req, res) => {
 	const { post_id } = req.params;
-	Post.findOne({ _id: post_id })
-		.then((post) => {
-			post.viewCnt++;
-			post.save((err) => {
-				if (err) return res.status(500).json({ error: err });
-				return res.status(200).json({ message: 
-					'post viewed successfully' 
+	if (post_id === undefined) {
+		return res.status(406).json({ message: 'parameter wrong' });
+	} else {
+		Post.findOne({ _id: post_id })
+			.then((post) => {
+				post.viewCnt++;
+				post.save((err) => {
+					if (err) return res.status(500).json({ error: err });
+					return res.status(200).json({ message: 
+						'post viewed successfully' 
+					});
 				});
+			})
+			.catch((err) => {
+				return res.status(500).json({ error: err });
 			});
-		})
-		.catch((err) => {
-			return res.status(500).json({ error: err });
-		});
+	}
 };
 
 exports.commentCreate = (req, res) => {
 	const { post_id, comment } = req.body;
-
-	User.findOne({ _id: req.decoded._id })
-		.then((user) => {
-			return user.username;
-		})
-		.then((username) => {
-			Post.findOne({ _id: post_id })
-				.then((post) => {
-					let content = {
-						comment: comment,
-						username: username
-					};
-					post.comments.push(content);
-					return post.save();
-				})
-				.then((post) => {
-					if (!post) return res.status(406).json({ message: 'no such post' });
-					return res.status(200).json({ message: 'comment successfully created' });
-				})
-				.catch((err) => {
-					return res.status(500).json({ error: err });
-				});
-		});
+	if (post_id === undefined || comment === undefined) {
+		return res.status(406).json({ message: 'parameter wrong' });
+	} else {
+		User.findOne({ _id: req.decoded._id })
+			.then((user) => {
+				return user.username;
+			})
+			.then((username) => {
+				Post.findOne({ _id: post_id })
+					.then((post) => {
+						let content = {
+							comment: comment,
+							username: username
+						};
+						post.comments.push(content);
+						return post.save();
+					})
+					.then((post) => {
+						if (!post) return res.status(406).json({ message: 'no such post' });
+						return res.status(200).json({ message: 'comment successfully created' });
+					})
+					.catch((err) => {
+						return res.status(500).json({ error: err });
+					});
+			});
+	}
 };
 
 exports.commentDelete = (req, res) => {
 	// const {  } = req.body;
 	const { post_id, comment_id, username } = req.params;
-	console.log(post_id, comment_id);
-	Post.findOne({ _id : post_id })
-		.then((post) => {
-			if (!post) return res.status(406).json({ message: 'no such post' });
-			for (let i=0;i<post.comments.length;i++) {
-				if (post.comments[i].username == username && post.comments[i]._id == comment_id) {
-					// console.log('yes');		
-					post.comments.splice(i, 1);
-					return post.save();
+	// console.log(post_id, comment_id);
+	if (post_id === undefined || comment_id === undefined || username === undefined) {
+		return res.status(406).json({ message: 'parameter wrong' });
+	} else {
+		Post.findOne({ _id : post_id })
+			.then((post) => {
+				if (!post) return res.status(406).json({ message: 'no such post' });
+				for (let i=0;i<post.comments.length;i++) {
+					if (post.comments[i].username == username && post.comments[i]._id == comment_id) {
+						// console.log('yes');		
+						post.comments.splice(i, 1);
+						return post.save();
+					}
 				}
-			}
-			console.log(post);
-		})
-		.then((post) => {
-			if (!post) return res.status(406).json({ message: 'no such post' });
-			return res.status(200).json({ message: 'comment successfully deleted' });
-		})
-		.catch((err) => {
-			return res.status(500).json({ error: err });
-		});
+				// console.log(post);
+			})
+			.then((post) => {
+				if (!post) return res.status(406).json({ message: 'no such post' });
+				return res.status(200).json({ message: 'comment successfully deleted' });
+			})
+			.catch((err) => {
+				return res.status(500).json({ error: err });
+			});
+	}
 };
